refactor: 데이터 클래스를 data 패키지로 이동
