+++ src/main/java/com/bookspot/batch/stock/file/StockFileDownloader.java
@@ -1,6 +1,7 @@
 package com.bookspot.batch.stock.file;
 
 import com.bookspot.batch.global.file.NaruFileDownloader;
+import com.bookspot.batch.stock.data.StockFileData;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
