+++ src/test/java/com/bookspot/batch/stock/file/StockFileInfoParserTest.java
@@ -2,6 +2,8 @@ package com.bookspot.batch.stock.file;
 
 import com.bookspot.batch.global.crawler.naru.CsvFilePath;
 import com.bookspot.batch.global.crawler.naru.NaruCrawler;
+import com.bookspot.batch.stock.data.CurrentLibrary;
+import com.bookspot.batch.stock.data.StockFileData;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
